third_party/blink/web_tests/platform/generic/shadow-dom/focus-slide-on-shadow-host-expected.txt

This is a testharness.js-based test.
PASS click on inner div should focus shadow host (with delegatesFocus = false).
FAIL click on inner div should focus inner focusable (with delegatesFocus = true). assert_equals: expected 1 but got 0
Harness: the test ran to completion.
